pyproject.toml
[project]
name = "document-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu<1.10.0",
    "langchain>=1.2.13",
    "langchain-mistralai>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.1",
    "networkx>=3.6.1",
    "nltk>=3.9.4",
    "numpy<2",
    "pdfplumber>=0.11.9",
    "protobuf>=7.34.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "sentencepiece>=0.2.1",
    "torch<2.3.0",
    "transformers<4.45.0",
]
